Create README.md
Password and Document Manager can store all our passwords and documents securely, so we don't have to worry about remembering and securing them. Password Manager allows us to use unique, strong passwords for all our important accounts rather than using the same password for all of them, which you should never do. In an increasingly digital world, safeguarding sensitive information is paramount. This mini project addresses this need by providing a comprehensive solution for secure password and document management on the Android platform. The project is totally built at user end and thus anyone can guarantee the access. The purpose of the project is to build an program to reduce the manual work for thinking about security of documents and creating strong password. This project report presents a secure password manager designed to protect digital credentials while ensuring user convenience. The manager employs encryption algorithms and access control mechanisms to safeguard passwords. The document management enables you to securely access and search Aadhar and PAN card documents securely stored on the server. We can able to download the documents using download option. This application keeps the user files safe and secured. The main purpose of this application is to provide security of the data. This application is very user friendly.